# project/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TEST API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/mqtt/data": {
      "get": {
        "tags": [
          "MQTT"
        ],
        "summary": "Get Latest Temperature Data",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "clientId": {
                  "type": "string",
                  "example": " "
                },
                "entity": {
                  "type": "string",
                  "example": " "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/mqtt/toggle": {
      "post": {
        "tags": [
          "MQTT"
        ],
        "summary": "Get Latest Temperature Data",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "clientId": {
                  "type": "string",
                  "example": " "
                },
                "entity": {
                  "type": "string",
                  "example": " "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/mqtt/automation": {
      "post": {
        "tags": [
          "Automation"
        ],
        "summary": "Set ON/OFF automation times for a device",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "type": "string",
                  "example": " "
                },
                "clientId": {
                  "type": "string",
                  "example": " "
                },
                "entity": {
                  "type": "string",
                  "example": " "
                },
                "onTime": {
                  "type": "string",
                  "example": " "
                },
                "offTime": {
                  "type": "string",
                  "example": " "
                },
                "timezone": {
                  "type": "string",
                  "example": "Asia/Ulaanbaatar"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/mqtt/update": {
      "put": {
        "tags": [
          "Update automation"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "ruleId": {
                  "type": "string",
                  "example": " "
                },
                "onTime": {
                  "type": "string",
                  "example": " "
                },
                "offTime": {
                  "type": "string",
                  "example": " "
                },
                "timezone": {
                  "type": "string",
                  "example": " "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/mqtt/getRule": {
      "get": {
        "tags": [
          "Get automation"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "clientId": {
                  "type": "string",
                  "example": " "
                },
                "entity": {
                  "type": "string",
                  "example": " "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/mqtt/delete": {
      "delete": {
        "tags": [
          "Delete automation"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/mqtt/powerlogs": {
      "get": {
        "tags": [
          "Power Logs"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "type": "string",
                  "example": " "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  }
}